ci: split tests (#1090)
